% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_marco.R
\name{analisis_marco}
\alias{analisis_marco}
\title{Obtiene los insumos para calcular las IF, cuando se trabaja con datos del marco}
\usage{
analisis_marco(base1, id, var1)
}
\arguments{
\item{base1}{base de datos a utilizar}

\item{id}{identificador de cada unidad}

\item{var1}{variable a analizar}
}
\value{
Permite obtener un data frame, con los descriptivos 
(n, suma, media, varianza, minimo, maximo) de la 
variable a analizar, ademas del folio a eliminar 
y su respectivo valor de la variable analizada. 
Cada fila es un calculo distinto: el primero incluye todas 
las unidades, el segundo excluye la unidad con mayor valor, 
el tercero las dos unidades con mayor valor, 
asi sucesivamente hasta llegar a n=1
}
\description{
Obtiene los insumos para calcular las IF, cuando se trabaja con datos del marco
}
\examples{
analisis_marco(base1 = LifeCycleSavings, id = sr, var1 = pop15)
}
