% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_mayor.R
\name{analisis_mayor}
\alias{analisis_mayor}
\title{Calcula las estimaciones eliminando la venta mayor de forma iterativa.}
\usage{
analisis_mayor(df_in, variable, folio, n = 1)
}
\arguments{
\item{df_in}{base de datos a utilizar}

\item{variable}{variable para la que deseo realizar los cálculos}

\item{folio}{identificador de cada unidad}

\item{n}{es una constante, igual a 1 (no modificar)}
}
\value{
una lista, el primer elemento es la base de datos completa y los siguientes son los descriptivos de la variable (el primer cálculo elimina la venta más grande, el segundo elimina las dos más grandes, así sucesivamente)
}
\description{
Calcula las estimaciones eliminando la venta mayor de forma iterativa.
}
\examples{
analisis_mayor(df_in = LifeCycleSavings, variable = pop15, folio = sr)
}
