% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_total.R
\name{analisis_total}
\alias{analisis_total}
\title{Title Calcula las estimaciones con todos los casos.}
\usage{
analisis_total(base, variable)
}
\arguments{
\item{base}{base de datos a utilizar}

\item{variable}{variable para la que deseo realizar los cálculos}
}
\value{
un objeto, con el número de casos, suma, media, varianza, mínimo y máximo de la variable escogida
}
\description{
Title Calcula las estimaciones con todos los casos.
}
\examples{
analisis_total(base = LifeCycleSavings, variable = pop15)
}
